quiz_api/requirements.txt
fastapi
uvicorn
python-dotenv
mistral-client